fix(retrieve): score使用f32，移除对嵌入向量不必要的Arc

+++ crates/soul-mem-algo/src/algo/retrieve/complex/assoc_with_action.rs
@@ -26,7 +26,7 @@ impl AssociateWithActionConfig {
     pub fn into_request(
         self,
         working_mem: Arc<WorkingMemory>,
-        source: Vec<(MemoryId, f64)>,
+        source: Vec<(MemoryId, f32)>,
     ) -> AssociateWithActionRequest {
         AssociateWithActionRequest {
             association: self
